Refactor earthquake detail components and utilities
- Replaced the SourceComparisonCompact component with CatalogComparison in EarthquakeDetailPage.
- Introduced useEarthquakeDetailViewModel hook to encapsulate earthquake data formatting logic.
- Enhanced local storage caching for earthquake details with versioning and timestamp.
- Updated SeismicWaveforms to allow toggling of station visibility with improved UI feedback.
- Added utility functions for formatting earthquake data, including magnitude, depth, coordinates, and event times.
- Created unit tests for new formatting utilities to ensure consistent behavior.
- Cleaned up EventMarker and SeismicWaveforms components by removing redundant code and improving readability.
