+++ src/components/CatalogComparison.css
