+++ src/pages/EarthquakeDetailPage.jsx
@@ -1,4 +1,4 @@
-import React, { lazy, Suspense, useCallback, useEffect, useMemo, useState } from 'react';
+import React, { lazy, Suspense, useEffect, useMemo, useState } from 'react';
 import { useLocation, useSearchParams } from 'react-router-dom';
 import axios from 'axios';
 import Header from '../components/Header';
@@ -6,14 +6,154 @@ import NearbyEvents from '../components/NearbyEvents';
 import LoadingScreen from '../components/LoadingScreen';
 import ErrorScreen from '../components/ErrorScreen';
 import moment from '../utils/time';
-import sanitizeHtml from '../utils/sanitizeHtml';
-import { generateEventSummary } from '../utils/generateEventSummary';
 import { backendHost } from '../utils/env';
-import { normalizeList } from '../utils/normalizeList';
-import { SourceComparisonCompact } from '../components/EarthquakeSourceComparison';
+import useEarthquakeDetailViewModel from '../hooks/useEarthquakeDetailViewModel';
+import CatalogComparison from '../components/CatalogComparison';
 import './EQInfoPage.css';
 
 const SeismicWaveforms = lazy(() => import('../components/SeismicWaveforms'));
+const DETAIL_CACHE_TTL_MS = 5 * 60 * 1000;
+const DETAIL_CACHE_VERSION = 1;
+const DETAIL_FETCH_TIMEOUT_MS = 12000;
+const RECENT_EVENTS_FETCH_TIMEOUT_MS = 20000;
+
+function getDetailCacheKey(eventId) {
+  return `earthquake-detail:${eventId}`;
+}
+
+function readCachedEarthquake(eventId) {
+  if (!eventId || typeof window === 'undefined') return null;
+
+  const cacheKey = getDetailCacheKey(eventId);
+  try {
+    const cached = window.localStorage.getItem(cacheKey);
+    if (!cached) return null;
+
+    const parsed = JSON.parse(cached);
+    if (parsed?.publicID === eventId && parsed.version !== DETAIL_CACHE_VERSION) {
+      writeCachedEarthquake(eventId, parsed);
+      return parsed;
+    }
+
+    const isFresh =
+      parsed?.version === DETAIL_CACHE_VERSION &&
+      parsed?.cachedAt &&
+      Date.now() - parsed.cachedAt < DETAIL_CACHE_TTL_MS;
+
+    if (!isFresh) {
+      window.localStorage.removeItem(cacheKey);
+      return null;
+    }
+
+    return parsed.payload || null;
+  } catch (_) {
+    window.localStorage.removeItem(cacheKey);
+    return null;
+  }
+}
+
+function writeCachedEarthquake(eventId, earthquake) {
+  if (!eventId || typeof window === 'undefined') return;
+
+  try {
+    window.localStorage.setItem(
+      getDetailCacheKey(eventId),
+      JSON.stringify({
+        cachedAt: Date.now(),
+        payload: earthquake,
+        version: DETAIL_CACHE_VERSION,
+      }),
+    );
+  } catch (_) {
+    // Ignore localStorage failures; fetched data is still usable for this render.
+  }
+}
+
+function useNearViewport(rootMargin = '600px') {
+  const [targetNode, setTargetNode] = useState(null);
+  const [isNearViewport, setIsNearViewport] = useState(false);
+
+  useEffect(() => {
+    if (isNearViewport || !targetNode) return undefined;
+    if (typeof IntersectionObserver === 'undefined') {
+      setIsNearViewport(true);
+      return undefined;
+    }
+
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        if (entry.isIntersecting) {
+          setIsNearViewport(true);
+          observer.disconnect();
+        }
+      },
+      { rootMargin },
+    );
+
+    observer.observe(targetNode);
+    return () => observer.disconnect();
+  }, [isNearViewport, rootMargin, targetNode]);
+
+  return [setTargetNode, isNearViewport];
+}
+
+function getEventRevisionMs(earthquake) {
+  const revisionTime =
+    earthquake?.last_modification || earthquake?.updated || earthquake?.modified || earthquake?.updatedAt;
+  const parsed = revisionTime ? new Date(revisionTime).getTime() : NaN;
+  return Number.isFinite(parsed) ? parsed : 0;
+}
+
+function isNewerEvent(candidate, current) {
+  const candidateRevision = getEventRevisionMs(candidate);
+  const currentRevision = getEventRevisionMs(current);
+  if (candidateRevision || currentRevision) {
+    return candidateRevision > currentRevision;
+  }
+  return JSON.stringify(candidate) !== JSON.stringify(current);
+}
+
+async function fetchEarthquakeByPublicID(eventId) {
+  const res = await axios.get(`${backendHost()}/eq-events/${encodeURIComponent(eventId)}`, {
+    timeout: DETAIL_FETCH_TIMEOUT_MS,
+    withCredentials: true,
+  });
+  return res.data?.payload || null;
+}
+
+async function fetchEarthquakeFromRecentEvents(eventId) {
+  const endDate = moment().format('YYYY-MM-DD HH:mm:ss');
+  const startDate = moment().subtract(90, 'days').format('YYYY-MM-DD HH:mm:ss');
+
+  const res = await axios.get(`${backendHost()}/eq-events`, {
+    params: { startTime: startDate, endTime: endDate },
+    timeout: RECENT_EVENTS_FETCH_TIMEOUT_MS,
+    withCredentials: true,
+  });
+
+  const events = res.data?.payload || [];
+  return events.find((ev) => ev.publicID === eventId) || null;
+}
+
+async function fetchEarthquakeWithFallback(eventId) {
+  try {
+    const directEvent = await fetchEarthquakeByPublicID(eventId);
+    if (directEvent) return directEvent;
+  } catch (_) {
+    // Fall back to the older range query so the frontend still works while
+    // the direct detail endpoint is being rolled out or restarted locally.
+  }
+
+  return fetchEarthquakeFromRecentEvents(eventId);
+}
+
+function getFetchErrorMessage(error) {
+  if (error?.code === 'ECONNABORTED') {
+    return 'The earthquake detail request timed out. Please check that the backend is running and try again.';
+  }
+
+  return error?.message || 'Failed to load earthquake data. Please try again or select from the events list.';
+}
 
 /**
  * Earthquake detail page for network-detected earthquakes. Displays event information
@@ -28,22 +168,29 @@ function EarthquakeDetailPage() {
   const [fetchedEarthquake, setFetchedEarthquake] = useState(null);
   const [isFetching, setIsFetching] = useState(false);
   const [fetchError, setFetchError] = useState(null);
+  const [waveformSentinelRef, shouldMountWaveforms] = useNearViewport();
   const eventId = searchParams.get('id');
   const isDevelopment =
     typeof process !== 'undefined' && process.env && process.env.NODE_ENV === 'development';
 
   const cachedEarthquake = useMemo(() => {
-    if (!eventId || typeof window === 'undefined') return null;
-    try {
-      const cached = window.localStorage.getItem(`earthquake-detail:${eventId}`);
-      return cached ? JSON.parse(cached) : null;
-    } catch (_) {
-      return null;
-    }
+    return readCachedEarthquake(eventId);
   }, [eventId]);
 
   // Get earthquake from navigation state, cached detail data, or a direct URL fetch.
   const earthquakeInfo = location.state?.earthquake || cachedEarthquake || fetchedEarthquake;
+  const {
+    coordText,
+    debugInfo,
+    depth,
+    formattedEventTimePh,
+    formattedEventTimeUtc,
+    formattedUpdatedTime,
+    formattedUpdatedTimeUtc,
+    pageTitle,
+    stationsForDisplay,
+    summaryMarkup,
+  } = useEarthquakeDetailViewModel(earthquakeInfo);
 
   // Fetch earthquake data if not available via location.state but ID is in URL
   useEffect(() => {
@@ -62,30 +209,13 @@ function EarthquakeDetailPage() {
       setIsFetching(true);
       setFetchError(null);
       try {
-        // Fetch events from a 90-day range to find the one with matching ID
-        // This covers most earthquake queries; adjust if needed
-        const endDate = moment().format('YYYY-MM-DD HH:mm:ss');
-        const startDate = moment().subtract(90, 'days').format('YYYY-MM-DD HH:mm:ss');
-
-        const res = await axios.get(`${backendHost()}/eq-events`, {
-          params: { startTime: startDate, endTime: endDate },
-          withCredentials: true,
-        });
-
-        const events = res.data?.payload || [];
-        const found = events.find((ev) => ev.publicID === eventId);
+        const found = await fetchEarthquakeWithFallback(eventId);
 
         if (!isMounted) return;
 
         if (found) {
           setFetchedEarthquake(found);
-          if (typeof window !== 'undefined') {
-            try {
-              window.localStorage.setItem(`earthquake-detail:${eventId}`, JSON.stringify(found));
-            } catch (_) {
-              // Ignore localStorage failures; fetched data is still usable for this render.
-            }
-          }
+          writeCachedEarthquake(eventId, found);
           setFetchError(null);
         } else {
           setFetchError(
@@ -98,9 +228,7 @@ function EarthquakeDetailPage() {
           // eslint-disable-next-line no-console
           console.error('Error fetching earthquake:', error);
         }
-        setFetchError(
-          error?.message || 'Failed to load earthquake data. Please try again or select from the events list.'
-        );
+        setFetchError(getFetchErrorMessage(error));
       } finally {
         if (isMounted) {
           setIsFetching(false);
@@ -115,100 +243,37 @@ function EarthquakeDetailPage() {
     };
   }, [cachedEarthquake, eventId, fetchedEarthquake, location.state]);
 
-  const formatEventTimePh = useCallback((eventTime) => {
-    const parsed = moment.utc(eventTime);
-    if (!parsed || !parsed.isValid()) return 'Date unavailable';
-    return parsed.add(8, 'hour').format('YYYY-MM-DD HH:mm:ss');
-  }, []);
+  useEffect(() => {
+    if (location.state?.earthquake || fetchedEarthquake || !cachedEarthquake || !eventId) {
+      return undefined;
+    }
 
-  const formatEventTimeUtc = useCallback((eventTime) => {
-    const parsed = moment.utc(eventTime);
-    if (!parsed || !parsed.isValid()) return 'Date unavailable';
-    return parsed.format('YYYY-MM-DD HH:mm:ss');
-  }, []);
+    let isMounted = true;
+    const revalidateCachedEarthquake = async () => {
+      try {
+        const freshEvent = await fetchEarthquakeByPublicID(eventId);
+        if (!isMounted || !freshEvent) return;
 
-  const formatCoord = useCallback((value, positiveLabel, negativeLabel) => {
-    const num = Number(value);
-    if (!Number.isFinite(num)) return null;
-    const hemi = num >= 0 ? positiveLabel : negativeLabel;
-    return `${Math.abs(num).toFixed(3)}°${hemi}`;
-  }, []);
+        if (isNewerEvent(freshEvent, cachedEarthquake)) {
+          setFetchedEarthquake(freshEvent);
+          writeCachedEarthquake(eventId, freshEvent);
+        }
+      } catch (_) {
+        // Keep the cached event when background revalidation fails.
+      }
+    };
 
-  const summaryMarkup = useMemo(
-    () => sanitizeHtml(earthquakeInfo?.eventSummary || generateEventSummary(earthquakeInfo)), //using generated summary as fallback if eventSummary is not provided
-    [earthquakeInfo],
-  );
-  const onlineStations = useMemo(
-    () => normalizeList(earthquakeInfo?.onlineStations),
-    [earthquakeInfo?.onlineStations],
-  );
+    revalidateCachedEarthquake();
 
-  // Network event detail uses online stations for waveform display.
-  const stationsForDisplay = useMemo(() => {
-    return onlineStations;
-  }, [onlineStations]);
+    return () => {
+      isMounted = false;
+    };
+  }, [cachedEarthquake, eventId, fetchedEarthquake, location.state]);
 
   // Debug: Check earthquake object structure
   if (isDevelopment && typeof window !== 'undefined') {
-    window._earthquakeDebug = {
-      hasEarthquakeInfo: !!earthquakeInfo,
-      earthquakeKeys: earthquakeInfo ? Object.keys(earthquakeInfo) : [],
-      onlineStations: onlineStations,
-      stationsForDisplay: stationsForDisplay,
-      onlineStationsLength: onlineStations.length,
-      stationsForDisplayLength: stationsForDisplay.length,
-    };
+    window._earthquakeDebug = debugInfo;
   }
-
-  const magnitude =
-    typeof earthquakeInfo?.magnitude === 'number'
-      ? earthquakeInfo.magnitude.toFixed(1).replace(/\.0$/, '')
-      : typeof earthquakeInfo?.magnitude_value === 'number'
-      ? earthquakeInfo.magnitude_value.toFixed(1).replace(/\.0$/, '')
-      : earthquakeInfo?.magnitude;
-
-  const depthValue = Number(
-    earthquakeInfo?.depth_km ?? earthquakeInfo?.depth ?? earthquakeInfo?.depth_value,
-  );
-  const depth = Number.isFinite(depthValue) ? `${depthValue.toFixed(0)} km` : null;
-
-  const eventTime = earthquakeInfo?.eventTime || earthquakeInfo?.OT;
-  const formattedEventTimePh = eventTime ? formatEventTimePh(eventTime) : null;
-  const formattedEventTimeUtc = eventTime ? formatEventTimeUtc(eventTime) : null;
-  const lastUpdatedTime =
-    earthquakeInfo?.last_modification || earthquakeInfo?.updated || earthquakeInfo?.modified || eventTime;
-  const formattedUpdatedTime = lastUpdatedTime ? formatEventTimePh(lastUpdatedTime) : null;
-  const formattedUpdatedTimeUtc = lastUpdatedTime ? formatEventTimeUtc(lastUpdatedTime) : null;
-
-  const latValue =
-    earthquakeInfo?.latitude_value ?? earthquakeInfo?.latitude ?? earthquakeInfo?.lat;
-  const lngValue =
-    earthquakeInfo?.longitude_value ?? earthquakeInfo?.longitude ?? earthquakeInfo?.lng;
-  const coordText = useMemo(() => {
-    const latText = formatCoord(latValue, 'N', 'S');
-    const lngText = formatCoord(lngValue, 'E', 'W');
-    if (!latText || !lngText) return null;
-    return `${latText}, ${lngText}`;
-  }, [formatCoord, latValue, lngValue]);
-
-  const placeDescription = earthquakeInfo?.place || '';
-  const genericLocation = earthquakeInfo?.location || earthquakeInfo?.text || 'Location unavailable';
-
-  // Use place description for location if available, otherwise use generic location
-  // Generate dynamic event title
-  const pageTitle = useMemo(() => {
-    if (earthquakeInfo?.title) return earthquakeInfo.title;
-    
-    const magText = magnitude ? `M${magnitude} Earthquake` : 'Earthquake';
-    
-    if (placeDescription && placeDescription !== 'Unavailable') {
-      return `${magText} ${placeDescription}`;
-    } else if (genericLocation) {
-      return `${magText} ${genericLocation}`;
-    }
-    
-    return magText;
-  }, [magnitude, placeDescription, genericLocation, earthquakeInfo?.title]);
 
   useEffect(() => {
     if (typeof document === 'undefined') return;
@@ -295,12 +360,18 @@ function EarthquakeDetailPage() {
           </section>
         )}
 
-        <Suspense fallback={<div className="eqinfo-panel muted">Loading waveform viewer...</div>}>
-          <SeismicWaveforms
-            earthquakeInfo={earthquakeInfo}
-            stations={stationsForDisplay}
-          />
-        </Suspense>
+        <div ref={waveformSentinelRef}>
+          {shouldMountWaveforms ? (
+            <Suspense fallback={<div className="eqinfo-panel muted">Loading waveform viewer...</div>}>
+              <SeismicWaveforms
+                earthquakeInfo={earthquakeInfo}
+                stations={stationsForDisplay}
+              />
+            </Suspense>
+          ) : (
+            <div className="eqinfo-panel muted">Preparing station recordings...</div>
+          )}
+        </div>
 
         <div className="eqinfo-grid">
           <div className="eqinfo-related-source-row">
@@ -310,7 +381,7 @@ function EarthquakeDetailPage() {
               distanceThresholdKm={200}
             />
 
-            <SourceComparisonCompact earthquakeInfo={earthquakeInfo} />
+            <CatalogComparison earthquakeInfo={earthquakeInfo} />
           </div>
         </div>
       </div>
